+++ resources/js/components/cadastros/treinamentoindustria/TreinamentoIndustria.vue
@@ -1,5 +1,11 @@
 <template>
     <div id="componenteTreinamentoIndustria">
+        <modal :modal-pai="modal" :titulo="titulo_janela_assinatura" :size="90" id="janelaAssinatura">
+            <template slot="conteudo">
+                <assinatura-carteira></assinatura-carteira>
+            </template>
+        </modal>
+
         <modal id="janelaCadastrar" :titulo="titulo_janela" :fechar="!preload" :size="90">
             <template slot="conteudo">
                 <preload v-show="preload"></preload>
@@ -130,6 +136,12 @@
                             data-target="#janelaCadastrar">
                         <i class="fa fa-plus"></i> Treinamento Indústria
                     </button>
+
+                    <button type="button" class="btn btn-sm btn-secondary"
+                            data-toggle="modal"
+                            data-target="#janelaAssinatura">
+                        <i class="fa fa-plus"></i> Assinatura Carteira
+                    </button>
                 </div>
             </form>
         </fieldset>
@@ -194,11 +206,13 @@
 <script>
 import controlePaginacao from '../../ControlePaginacao';
 import modal from '../../Modal';
+import AssinaturaCarteira from './AssinaturaCarteira.vue';
 
 export default {
     components: {
         modal,
         controlePaginacao,
+        AssinaturaCarteira,
     },
     props: {
         qntPag: {
@@ -226,6 +240,7 @@ export default {
         return {
             hash: String(Math.random()).substr(2),
             titulo_janela: '',
+            titulo_janela_assinatura: 'Assinatura Carteira',
 
             preload: false,
             editando: false,